requirements.txt

# Core
python-dotenv>=1.0.0

# Streamlit UI
streamlit>=1.32.0

# LLM - Mistral
mistralai>=0.4.2

# Embeddings & RAG
langchain>=0.1.16
langchain-community>=0.0.32
sentence-transformers>=2.2.2

# Vector DB
pinecone==3.0.0

# Utils
pydantic>=2.5.0
